+++ api/utils.go
@@ -1,6 +1,10 @@
 package api
 
-import "time"
+import (
+	"strconv"
+	"strings"
+	"time"
+)
 
 // boolToPtr returns the pointer to a boolean
 func boolToPtr(b bool) *bool {
@@ -25,4 +29,25 @@ func stringToPtr(str string) *string {
 // timeToPtr returns the pointer to a time stamp
 func timeToPtr(t time.Duration) *time.Duration {
 	return &t
+}
+
+// formatFloat converts the floating-point number f to a string,
+// after rounding it to the passed unit.
+//
+// Uses 'f' format (-ddd.dddddd, no exponent), and uses at most
+// maxPrec digits after the decimal point.
+func formatFloat(f float64, maxPrec int) string {
+	v := strconv.FormatFloat(f, 'f', -1, 64)
+
+	idx := strings.LastIndex(v, ".")
+	if idx == -1 {
+		return v
+	}
+
+	sublen := idx + maxPrec + 1
+	if sublen > len(v) {
+		sublen = len(v)
+	}
+
+	return v[:sublen]
 }
